Update README.md
PROJECT OVERVIEW
A Gym Management System is a comprehensive solution designed to streamline and automate the operations of a gym or fitness center. This project serves as an all-in-one platform for managing memberships, scheduling classes, tracking payments, and maintaining member information. Built using Flask for backend functionality and HTML/CSS for the user interface, the system ensures an efficient and user-friendly experience. Key features include a dynamic dashboard for administrators, secure login for members, and tools for tracking attendance and fitness goals. This project aims to enhance operational efficiency and provide a seamless experience for both gym staff and members.